wait for job result before emitting complete. (#85)
Fixes #8 

**This would be a breaking change**

Currently, if you call `table.createWriteStream()`, the stream will emit "complete" with a Job object. You can then use that Job object to assign event handlers, as usual with an LRO. Judging from #8, I'm not sure this pattern is clear to the users. And aside from that, I'm not sure what we're doing is ideal, either.

Ideally, we would not emit "complete" until the Job finished, instead of making the user do it manually. This PR shows how easy it would be for us to implement. However, I *believe* this would be a breaking change, although it's hard to say if it's actually a bugfix.

I'm leaving this open for discussion on how we should handle this. Please share your thoughts!

---

### Before
```js
table.createWriteStream(...)
  .on('complete', job => {
    job
      .on('error', err => {
        // Job failed.
      })
      .on('complete', () => {
        // Job completed successfully.
      });
  });
```

### After
```js
table.createWriteStream(...)
  .on('error', err => {
    // Job failed.
  })
  .on('job', job => {
    // `job` is a Job object.
  })
  .on('finish', () => {
    // Job completed successfully.
  });
```

And, `complete` still works to combine those. I added the "finish" event, because that's the natural writable stream closing event.

```js
table.createWriteStream(...)
  .on('error', err => {
    // Job failed.
  })
  .on('complete', job => {
    // `job` is a Job object.
    // Job completed successfully.
  });
```