build: fix path-to-regexp to older version due to node 14 requirement (#1403)
Sinon depends on the [`nise` package](https://www.npmjs.com/package/nise). That package just published `v6.0.1` which updated `path-to-regexp` from `v6.2.1` to `v8.1.0` with https://togithub.com/sinonjs/nise/pull/226. That should have been a major version bump, as this ended up introducing a breaking change that now it depends on `Node >= 16`, which is causing our CI to break. 

This PR fixes `path-to-regexp` to `^6.2.1` like previously to make it work on `Node >= 14`

Similar to https://togithub.com/googleapis/nodejs-bigquery-storage/pull/475